Add escapeNestedDots option for preserving key structure.
This module did not previously offer a way to encode any nested '.'
characters in a key such that it could be easily identified for use by
the end user or other modules. As a result, any keys which had nested
'.' characters appeared as though they were simply another nested object
layer, but this could cause issues if trying to reconstruct the object.
This commit adds the escapeNestedDots option which allows the user to
optionally specify that this information should be encoded for later use
by the consuming software.

Related to https://github.com/mrodrig/json-2-csv/issues/184